[TH2-2870] Moved queue logic into routers
