+++ th2_common/schema/message/impl/rabbitmq/raw/rabbit_raw_batch_router.py
@@ -14,12 +14,16 @@
 
 from th2_grpc_common.common_pb2 import RawMessageBatch, RawMessage
 
+from th2_common.schema.filter.strategy.impl.default_filter_strategy import DefaultFilterStrategy
 from th2_common.schema.message.configuration.message_configuration import QueueConfiguration
+from th2_common.schema.message.impl.rabbitmq.configuration.subscribe_target import SubscribeTarget
 from th2_common.schema.message.impl.rabbitmq.connection.connection_manager import ConnectionManager
-from th2_common.schema.message.impl.rabbitmq.raw.rabbit_raw_batch_queue import RabbitRawBatchQueue
+from th2_common.schema.message.impl.rabbitmq.raw.rabbit_raw_batch_sender import RabbitRawBatchSender
+from th2_common.schema.message.impl.rabbitmq.raw.rabbit_raw_batch_subscriber import RabbitRawBatchSubscriber
 from th2_common.schema.message.impl.rabbitmq.router.abstract_rabbit_batch_message_router import \
     AbstractRabbitBatchMessageRouter
-from th2_common.schema.message.message_queue import MessageQueue
+from th2_common.schema.message.message_sender import MessageSender
+from th2_common.schema.message.message_subscriber import MessageSubscriber
 from th2_common.schema.message.queue_attribute import QueueAttribute
 
 
@@ -42,6 +46,14 @@ class RabbitRawBatchRouter(AbstractRabbitBatchMessageRouter):
     def _add_message(self, batch: RawMessageBatch, message: RawMessage):
         batch.messages.append(message)
 
-    def _create_queue(self, connection_manager: ConnectionManager,
-                      queue_configuration: QueueConfiguration) -> MessageQueue:
-        return RabbitRawBatchQueue(connection_manager, queue_configuration)
+    def create_sender(self, connection_manager: ConnectionManager,
+                      queue_configuration: QueueConfiguration) -> MessageSender:
+        return RabbitRawBatchSender(connection_manager, queue_configuration.exchange, queue_configuration.routing_key)
+
+    def create_subscriber(self, connection_manager: ConnectionManager,
+                          queue_configuration: QueueConfiguration) -> MessageSubscriber:
+        subscribe_target = SubscribeTarget(queue_configuration.queue, queue_configuration.routing_key)
+        return RabbitRawBatchSubscriber(connection_manager,
+                                        queue_configuration,
+                                        DefaultFilterStrategy(),
+                                        subscribe_target)
